Move gpio code into single file
