+++ CMakeLists_bsp_stg.txt
@@ -41,6 +41,7 @@ set(BSP_SOURCES
   ${BSP_DIR}/common/include/bsp_mx.h
   ${BSP_DIR}/common/include/bsp_pinout.h
   ${BSP_DIR}/common/include/bsp_qpc.h
+  ${BSP_DIR}/common/src/gpio.c
   ${BSP_DIR}/common/src/i2c.c
   ${BSP_DIR}/common/src/iwdg.c
   ${BSP_DIR}/common/src/rtc.c
@@ -56,5 +57,4 @@ set(BSP_SOURCES
 # model specific
 list(APPEND BSP_SOURCES
   ${BSP_DIR}/${STG_MODEL}/src/adc.c
-  ${BSP_DIR}/${STG_MODEL}/src/gpio.c
 )
